refactor file structure and __init__
