+++ src/main/python/pybuilder_aws_plugin/__init__.py
@@ -1,20 +1,20 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# flake8: noqa
 
-from pybuilder.core import init
 import sys
 
-from .upload_zip_task import upload_zip_to_s3, package_lambda_code
-upload_zip_to_s3  # Linting purposes, no other use
-package_lambda_code  # Linting purposes, no other use
+from pybuilder.core import init
+
+from .lambda_tasks import upload_zip_to_s3, package_lambda_code
 
 if sys.version_info[0:2] >= (2, 7):
-    from .upload_cfn_task import upload_cfn_to_s3
-    upload_cfn_to_s3  # Linting purposes, no other use
+    from .cfn_tasks import upload_cfn_to_s3
 
 
 @init
 def initialize_plugin(project):
+    """ Setup plugin defaults. """
     project.set_property(
         'lambda_file_access_control', 'bucket-owner-full-control')
     project.set_property('bucket_prefix', '')
